Make pattern of defining and registering configs consistent between Kibana and X-Pack.
