+++ src/ui/public/management/index_pattern_creation/register.js
